% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{AddLabel}
\alias{AddLabel}
\title{Add annotate}
\usage{
AddLabel(
  textLabelx = NULL,
  textLabely = NULL,
  textLabel = "text label",
  textLabelColor = NULL,
  textLabelSize = NULL,
  textLabelAngle = NULL,
  textLabelFontface = NULL
)
}
\arguments{
\item{textLabelx}{annotate x position}

\item{textLabely}{annotate y position}

\item{textLabel}{annotate label text}

\item{textLabelColor}{annotate color}

\item{textLabelSize}{annotate size}

\item{textLabelAngle}{annotate angle}

\item{textLabelFontface}{annotate font face}
}
\value{
A ggolot2 object
}
\description{
Add annotate
}
