% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{addLine}
\alias{addLine}
\title{Add line}
\usage{
addLine(
  xstart,
  ystart,
  xend,
  yend,
  color = "black",
  lineType = 1,
  lineWidth = 1
)
}
\arguments{
\item{xstart}{x start point}

\item{ystart}{y start point}

\item{xend}{x end point}

\item{yend}{y end point}

\item{color}{color of line}

\item{lineWidth}{linee width}

\item{linetype}{line type}
}
\value{
A ggplot2 object
}
\description{
Add line
}
