% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{CaptionTheme}
\alias{CaptionTheme}
\title{Set Caption theme}
\usage{
CaptionTheme(
  ploCaptionFace = NULL,
  ploCaptionColor = NULL,
  ploCaptionSize = NULL,
  ploCaptionHjust = NULL,
  ploCaptionVjust = NULL,
  ploCaptionAngle = NULL
)
}
\arguments{
\item{ploCaptionFace}{caption face}

\item{ploCaptionColor}{caption color}

\item{ploCaptionSize}{caption size}

\item{ploCaptionHjust}{caption horizontal position}

\item{ploCaptionVjust}{caption vertical position}

\item{ploCaptionAngle}{caption angle}
}
\value{
A ggolot2 object
}
\description{
Set Caption theme
}
