% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterdata.R
\name{classificationModelEva}
\alias{classificationModelEva}
\title{Classificatio model evaluation function}
\usage{
classificationModelEva(truth, estimate)
}
\arguments{
\item{truth}{lable of data}

\item{estimate}{prediction of model}
}
\value{
a list of result
}
\description{
Classificatio model evaluation function
}
\examples{
data("two_class_example")
classificationModelEva(two_class_example$truth,two_class_example$predicted)
}
