% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_r_Z.R
\name{convert_r_Z}
\alias{convert_r_Z}
\title{Convert a Correlation Coefficient (\emph{r}) Into the Scale of Fisher's \emph{Z}}
\usage{
convert_r_Z(r)
}
\arguments{
\item{r}{Correlation coefficient (between two variables)}
}
\value{
A 1-row \code{data.frame} with columns \code{term} and
  \code{value}. The \code{term} is \code{"Z_from_r"} and \code{value}
  is Fisher's \emph{Z} corresponding to the supplied correlation
  coefficient. The inverse direction is \code{\link{convert_Z_r}}.
}
\description{
This function converts a correlation coefficient into the scale of
Fisher's \emph{Z}, the variance-stabilizing transformation of a
correlation. Many authors call this map the \emph{z}-prime transform
and write the transformed value as \emph{z}'. The capital \emph{Z} is
meaningful: Fisher's \emph{Z} is not a \emph{z}-score (it is not a
standardized variate, that is, an observation centered and divided by a
standard deviation). It is the transform \eqn{Z = \mathrm{atanh}(r)} of
a correlation coefficient, applied because the sampling distribution of
\emph{Z} is approximately normal with a variance that does not depend on
the population correlation, which makes \emph{Z} convenient for forming
confidence intervals.
}
\details{
This function is typically used in the context of forming a confidence interval for a population correlation coefficient. Note that, in that situation, the two variables are assumed to follow a bivariate normal distribution (e.g., Hays, 1994).
}
\examples{
# From Hays (1994, pp. 649--650)
convert_r_Z(.35)


}
\references{
Kelley, K. (2007). Confidence intervals for standardized effect sizes: Theory, application, and implementation. \emph{Journal of Statistical Software, 20}(8), 1--24.
  \doi{10.18637/jss.v020.i08}

Hays, W. L. (1994). \emph{Statistics} (5th ed.). Fort Worth, TX:
  Harcourt Brace College Publishers.
}
\seealso{
\code{\link{convert_Z_r}}, \code{\link{ci_r}}

Other parameterization conversions:
\code{\link[=convert_F_chisq]{convert_F_chisq()}},
\code{\link{convert_R2}},
\code{\link[=convert_Z_r]{convert_Z_r()}},
\code{\link[=convert_cor_cov]{convert_cor_cov()}},
\code{\link[=convert_d_or]{convert_d_or()}},
\code{\link[=convert_d_r]{convert_d_r()}},
\code{\link{convert_t_smd}},
\code{\link[=convert_z_normal]{convert_z_normal()}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\concept{parameterization conversions}
\keyword{design}
