% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_t_smd.R
\name{convert_t_smd}
\alias{convert_t_smd}
\alias{convert_delta_lambda}
\alias{convert_lambda_delta}
\title{Conversion Functions for Noncentral \emph{t}-distribution}
\usage{
convert_delta_lambda(delta, n_1, n_2)

convert_lambda_delta(lambda, n_1, n_2)
}
\arguments{
\item{delta}{Population value of the standardized mean difference}

\item{n_1}{Sample size in group 1}

\item{n_2}{Sample size in group 2}

\item{lambda}{noncentral value from a \emph{t}-distribution}
}
\value{
Each function returns a 1-row \code{data.frame} with columns
  \code{term} and \code{value}. The \code{term} entry identifies the
  conversion (\code{"delta_lambda"} or \code{"lambda_delta"}) and
  \code{value} is the converted scalar. The two functions are exact
  inverses given the \emph{per-group} sample sizes
  \code{n_1} and \code{n_2}.
}
\description{
Functions useful for converting a standardized mean difference to a noncentrality parameter, and vice versa.
}
\details{
Although \code{lambda} is the population noncentral value, an estimate of it is the observed value of a
\emph{t}-statistic. Likewise, delta can be estimated as the observed standardized mean difference. Thus, the observed
standardized mean difference can be converted to the observed \emph{t}-value. These functions are especially helpful in the
context of forming confidence intervals for the population standardized mean difference.
}
\examples{
convert_lambda_delta(lambda = 2, n_1 = 113, n_2 = 113)
convert_delta_lambda(delta = .266076, n_1 = 113, n_2 = 113)

}
\references{
Kelley, K. (2007). Confidence intervals for standardized effect sizes:
Theory, application, and implementation. \emph{Journal of Statistical
Software, 20}(8), 1--24. \doi{10.18637/jss.v020.i08}
}
\seealso{
\code{\link{smd}}, \code{\link{ci_smd}}, \code{\link{ss_aipe_smd}}

Other parameterization conversions:
\code{\link[=convert_F_chisq]{convert_F_chisq()}},
\code{\link{convert_R2}},
\code{\link[=convert_Z_r]{convert_Z_r()}},
\code{\link[=convert_cor_cov]{convert_cor_cov()}},
\code{\link[=convert_d_or]{convert_d_or()}},
\code{\link[=convert_d_r]{convert_d_r()}},
\code{\link[=convert_r_Z]{convert_r_Z()}},
\code{\link[=convert_z_normal]{convert_z_normal()}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\concept{parameterization conversions}
\keyword{design}
