% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Z_r.R
\name{convert_Z_r}
\alias{convert_Z_r}
\title{Convert Fisher's \emph{Z} Into the Scale of a Correlation Coefficient (\emph{r})}
\usage{
convert_Z_r(Z)
}
\arguments{
\item{Z}{Fisher's \emph{Z} (the variance-stabilizing transform of a
correlation, which many authors call \emph{z}')}
}
\value{
A 1-row \code{data.frame} with columns \code{term} and
  \code{value}. The \code{term} is \code{"r_from_Z"} and \code{value}
  is the correlation coefficient corresponding to the supplied
  Fisher's \emph{Z}. The inverse direction is
  \code{\link{convert_r_Z}}.
}
\description{
Converts Fisher's \emph{Z} back into the scale of a correlation
coefficient (\emph{r}). Fisher's \emph{Z} is the variance-stabilizing
transformation of a correlation; many authors call it the \emph{z}-prime
transform and write the transformed value as \emph{z}'. The capital
\emph{Z} is meaningful: Fisher's \emph{Z} is not a \emph{z}-score (it is
not a standardized variate, that is, an observation centered and divided
by a standard deviation). This function applies the inverse transform
\eqn{r = \mathrm{tanh}(Z)} to return to the scale of a correlation
coefficient.
}
\details{
This function is typically used in the context of forming a confidence interval for a population correlation coefficient. Note that, in that situation, the two variables are assumed to follow a bivariate normal distribution (e.g., Hays, 1994).
}
\examples{
# From Hays (1994, pp. 649--650)
convert_Z_r(0.3654438)


}
\references{
Kelley, K. (2007). Confidence intervals for standardized effect sizes: Theory, application, and implementation. \emph{Journal of Statistical Software, 20}(8), 1--24.
  \doi{10.18637/jss.v020.i08}

Hays, W. L. (1994). \emph{Statistics} (5th ed.). Fort Worth, TX:
  Harcourt Brace College Publishers.
}
\seealso{
\code{\link{convert_r_Z}}, \code{\link{ci_r}}

Other parameterization conversions:
\code{\link[=convert_F_chisq]{convert_F_chisq()}},
\code{\link{convert_R2}},
\code{\link[=convert_cor_cov]{convert_cor_cov()}},
\code{\link[=convert_d_or]{convert_d_or()}},
\code{\link[=convert_d_r]{convert_d_r()}},
\code{\link[=convert_r_Z]{convert_r_Z()}},
\code{\link{convert_t_smd}},
\code{\link[=convert_z_normal]{convert_z_normal()}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\concept{parameterization conversions}
\keyword{design}
