% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_t.R
\name{cv_t}
\alias{cv_t}
\title{Provides the Critical Value(s) for a \emph{t}-distribution}
\usage{
cv_t(
  alpha_level,
  df,
  alternative = "not_equal",
  alpha_lower,
  alpha_upper,
  ncp = 0,
  verbose = TRUE
)
}
\arguments{
\item{alpha_level}{Type I error rate (i.e., the false positive rate).}

\item{df}{The number of degrees of freedom (a positive number)}

\item{alternative}{The type of alternative hypothesis of interest.}

\item{alpha_lower}{The error rate on the lower (negative) side of the distribution.}

\item{alpha_upper}{The error rate on the upper (positive) side of the distribution.}

\item{ncp}{The noncentral parameter (if zero, the default, it is the central \emph{t}-distribution).}

\item{verbose}{Provides extra information about areas under the curve.}
}
\value{
Returns the critical value(s), based on the input specifications, in a output style.
}
\description{
Provides the Critical Value(s) for a \emph{t}-distribution
}
\details{
Though a noncentral parameter can be included, that would not be done for a standard null hypothesis significance test.
}
\examples{
# A basic call for finding critical values with equal area in the two tails.
cv_t(alpha_level = .05, df = 13)

# A basic call for a single-sided confidence interval (for "a greater than" alternative hypothesis)
cv_t(alpha_level = .05, df = 13, alternative = "greater")

# A single-sided confidence interval (for "a greater than" alternative hypothesis); simple output.
cv_t(alpha_lower = 0, alpha_upper = .05, df = 13, verbose = FALSE)

# For a nonsymmetric 95\% confidence interval.
cv_t(alpha_lower = .01, alpha_upper = .04, df = 13)

}
\references{
Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing
  experiments and analyzing data: A model comparison perspective}
  (4th ed.). Routledge.
}
\seealso{
Other critical values:
\code{\link[=cv_bonferroni_f]{cv_bonferroni_f()}},
\code{\link[=cv_bryant_paulson]{cv_bryant_paulson()}},
\code{\link[=cv_chisq]{cv_chisq()}},
\code{\link[=cv_dunnett]{cv_dunnett()}},
\code{\link[=cv_f]{cv_f()}},
\code{\link[=cv_scheffe]{cv_scheffe()}},
\code{\link[=cv_smm]{cv_smm()}},
\code{\link[=cv_tukey_hsd]{cv_tukey_hsd()}},
\code{\link[=cv_z]{cv_z()}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\concept{critical values}
