% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_z.R
\name{cv_z}
\alias{cv_z}
\title{Provides the Critical Value(s) for the Standard Normal Distribution (the \emph{z}-distribution, With Mean 0 and Variance 1)}
\usage{
cv_z(
  alpha_level,
  alternative = "not_equal",
  alpha_lower,
  alpha_upper,
  verbose = TRUE
)
}
\arguments{
\item{alpha_level}{Type I error rate (i.e., the false positive rate).}

\item{alternative}{The type of alternative hypothesis of interest.}

\item{alpha_lower}{The error rate on the lower (negative) side of the distribution.}

\item{alpha_upper}{The error rate on the upper (positive) side of the distribution.}

\item{verbose}{Provides extra information about areas under the curve.}
}
\value{
Returns the critical value(s), based on the input specifications, in a output style.
}
\description{
Provides the Critical Value(s) for the Standard Normal Distribution (the \emph{z}-distribution, With Mean 0 and Variance 1)
}
\examples{
# A basic call for finding critical values with equal area in the two tails.
cv_z(alpha_level = .05)

# A basic call for a single-sided confidence interval (for "a greater than" alternative hypothesis)
cv_z(alpha_level = .05, alternative = "greater")

# A single-sided confidence interval (for "a greater than" alternative hypothesis); simple output.
cv_z(alpha_lower = 0, alpha_upper = .05, verbose = FALSE)

# For a nonsymmetric 95\% confidence interval.
cv_z(alpha_lower = .01, alpha_upper = .04)

}
\seealso{
Other critical values:
\code{\link[=cv_bonferroni_f]{cv_bonferroni_f()}},
\code{\link[=cv_bryant_paulson]{cv_bryant_paulson()}},
\code{\link[=cv_chisq]{cv_chisq()}},
\code{\link[=cv_dunnett]{cv_dunnett()}},
\code{\link[=cv_f]{cv_f()}},
\code{\link[=cv_scheffe]{cv_scheffe()}},
\code{\link[=cv_smm]{cv_smm()}},
\code{\link[=cv_t]{cv_t()}},
\code{\link[=cv_tukey_hsd]{cv_tukey_hsd()}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\concept{critical values}
