% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_R2.R
\name{expected_R2}
\alias{expected_R2}
\title{Expected Value of the Squared Multiple Correlation Coefficient}
\usage{
expected_R2(population_R2, N, p)
}
\arguments{
\item{population_R2}{Population squared multiple correlation coefficient}

\item{N}{Sample size}

\item{p}{The number of predictor variables}
}
\value{
A 1-row \code{data.frame} with columns \code{term} and \code{value}.
The \code{term} value is \code{"expected_value_population_R2"} and
\code{value} is the expected value of \eqn{R^2} under random
sampling.
}
\description{
Computes the expected value of the observed squared multiple correlation
coefficient given the population squared multiple correlation coefficient,
the sample size, and the number of predictors. The sample \eqn{R^2} is a
positively biased estimator of its population value, and the expected value
quantifies how large that bias is for a particular design.
}
\details{
Uses the hypergeometric function as discussed in section 28 of Stuart, Ord, and Arnold (1999) in order to obtain the \emph{correct}
value for the squared multiple correlation coefficient. Many times an exact value is given that ignores the hypergeometric function.
This function yields the correct value.
}
\examples{
expected_R2(.5, 10, 5)
expected_R2(.5, 25, 5)
expected_R2(.5, 50, 5)
expected_R2(.5, 100, 5)
expected_R2(.5, 1000, 5)
expected_R2(.5, 10000, 5)

}
\references{
Kelley, K. (2008). Sample size planning for the squared multiple
  correlation coefficient: Accuracy in parameter estimation via narrow
  confidence intervals. \emph{Multivariate Behavioral Research, 43},
  524--555. \doi{10.1080/00273170802490632}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing
  experiments and analyzing data: A model comparison perspective}
  (4th ed.). Routledge. (See Chapter 3 on \eqn{R^2} as a model comparison
  effect size.)

Olkin, I., & Pratt, J. W. (1958). Unbiased estimation of certain
  correlation coefficients. \emph{The Annals of Mathematical Statistics,
  29}(1), 201--211.

Stuart, A., Ord, J. K., & Arnold, S. (1999). \emph{Kendall's advanced
  theory of statistics, volume 2A: Classical inference and the linear
  model} (6th ed.). Arnold.
}
\seealso{
\code{\link{ss_aipe_R2}}, \code{\link{ci_R2}}, \code{\link{var_R2}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
