% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmr.R
\name{glance.mlmr}
\alias{glance.mlmr}
\title{Glance at an Mlmr Fit}
\usage{
\method{glance}{mlmr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"mlmr"}.}

\item{\dots}{Unused.}
}
\value{
A one-row \code{data.frame}.
}
\description{
Returns a one-row \code{data.frame} of model-level summaries in the
column convention used by the \pkg{broom} ecosystem
(\code{R2}, \code{adj_R2}, \code{sigma},
\code{statistic}, \code{p_value}, \code{df}, \code{logLik},
\code{AIC}, \code{BIC}, \code{deviance}, \code{df_residual},
\code{nobs}). The \code{statistic} and \code{p_value} columns
report the omnibus likelihood ratio test of all slopes equal to
zero (the FIML analog of the \code{lm} omnibus \emph{F}-test).
}
\examples{
fit <- mlmr(t6_paragraph_comprehension ~ t5_general_information +
              t9_word_meaning,
            data = holzinger_swineford, ci_method = "wald")
generics::glance(fit)

}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
