% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{guideColorLabelName}
\alias{guideColorLabelName}
\title{Change Color guide name and color}
\usage{
guideColorLabelName(
  guideLabelName = c("a", "b"),
  guideLabelcolor = c("red", "yellow")
)
}
\arguments{
\item{guideLabelName}{legend label text}

\item{guideLabelcolor}{legend label color}
}
\value{
A ggolot2 object
}
\description{
Change Color guide name and color
}
