% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{guideFillLabelName}
\alias{guideFillLabelName}
\title{Change fill guide name and color}
\usage{
guideFillLabelName(
  guideLabelName = c("a", "b"),
  guideLabelcolor = c("red", "yellow")
)
}
\arguments{
\item{guideLabelName}{legend label text}

\item{guideLabelcolor}{legend label color}
}
\value{
A ggolot2 object
'
}
\description{
Change fill guide name and color
}
