% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{guideFillSetting}
\alias{guideFillSetting}
\title{fill guide}
\usage{
guideFillSetting(
  guideTitle = "图例标题",
  guideTitleFace = NULL,
  guideTitleColor = NULL,
  guideTitleSize = NULL,
  guideTitleHjust = NULL,
  guideTitleVjust = NULL,
  guideTitleAngle = NULL,
  guideLabelFace = NULL,
  guideLabelColor = NULL,
  guideLabelSize = NULL,
  guideLabelHjust = NULL,
  guideLabelVjust = NULL,
  guideLabelAngle = NULL
)
}
\arguments{
\item{guideTitle}{legend title}

\item{guideTitleFace}{legend face}

\item{guideTitleColor}{legend color}

\item{guideTitleSize}{legend size}

\item{guideTitleHjust}{legend x positon}

\item{guideTitleVjust}{legend y position}

\item{guideTitleAngle}{legend angle}

\item{guideLabelFace}{legend label face}

\item{guideLabelColor}{legend lable color}

\item{guideLabelSize}{legend label size}

\item{guideLabelHjust}{legend label x postion}

\item{guideLabelVjust}{legend label y position}

\item{guideLabelAngle}{legend label angle}
}
\value{
A ggolot2 object
}
\description{
fill guide
}
