% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{GuildFillSetting}
\alias{GuildFillSetting}
\title{fill guide}
\usage{
GuildFillSetting(
  guildTitle = "图例标题",
  guildTitleFace = NULL,
  guildTitleFamily = NULL,
  guildTitleColor = NULL,
  guildTitleSize = NULL,
  guildTitleHjust = NULL,
  guildTitleVjust = NULL,
  guildTitleAngle = NULL,
  guildLabelFace = NULL,
  guildLabelFamily = NULL,
  guildLabelColor = NULL,
  guildLabelSize = NULL,
  guildLabelHjust = NULL,
  guildLabelVjust = NULL,
  guildLabelAngle = NULL
)
}
\arguments{
\item{guildTitleFamily}{legend font family}

\item{guildLabelFamily}{legend label font family}

\item{guideTitle}{legend title}

\item{guideTitleFace}{legend font face}

\item{guideTitleColor}{legend color}

\item{guideTitleSize}{legend size}

\item{guideTitleHjust}{legend x positon}

\item{guideTitleVjust}{legend y position}

\item{guideTitleAngle}{legend angle}

\item{guideLabelFace}{legend label font face}

\item{guideLabelColor}{legend lable color}

\item{guideLabelSize}{legend label size}

\item{guideLabelHjust}{legend label x postion}

\item{guideLabelVjust}{legend label y position}

\item{guideLabelAngle}{legend label angle}
}
\value{
A ggolot2 object
}
\description{
fill guide
}
