% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{legendPosition}
\alias{legendPosition}
\title{Change legend position}
\usage{
legendPosition(legend.position = "left")
}
\arguments{
\item{legend.position}{Position of legend. option left, right, top, bottom or two element numeric vector}
}
\value{
A ggplot2 object
}
\description{
Change legend position
}
