% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterdata.R
\name{regModelEva}
\alias{regModelEva}
\title{Regression model evaluation function}
\usage{
regModelEva(truth, estimate)
}
\arguments{
\item{truth}{lable of data}

\item{estimate}{prediction of model}
}
\value{
a list of result
}
\description{
Regression model evaluation function
}
\examples{
data("solubility_test")
regModelEva(solubility_test$solubility,solubility_test$prediction)
}
