% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterdata.R
\name{scaleData}
\alias{scaleData}
\title{Scale data}
\usage{
scaleData(data, col, all = F, center = T, scale = T)
}
\arguments{
\item{data}{dataset}

\item{col}{column of dataset}

\item{all}{if all=F, remove original column, otherwise keep}

\item{center}{logical value or numeric vector of length equal to the number of columns}

\item{scale}{logical value or numeric vector of length equal to the number of columns}
}
\value{
scale dataset
}
\description{
Scale data
}
\examples{
# scale data to 0-1
maxs <- apply(iris[,c(1,2)], 2, max)
mins <- apply(iris[,c(1,2)], 2, min)
#scaleData(iris,col = c(1,2),all = T,center = mins,scale = maxs-mins)
}
