% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd}
\alias{smd}
\title{Standardized Mean Difference}
\usage{
smd(
  group_1 = NULL,
  group_2 = NULL,
  mean_1 = NULL,
  mean_2 = NULL,
  s_1 = NULL,
  s_2 = NULL,
  s = NULL,
  n_1 = NULL,
  n_2 = NULL,
  unbiased = FALSE
)
}
\arguments{
\item{group_1}{Raw data for group 1}

\item{group_2}{Raw data for group 2}

\item{mean_1}{The mean of group 1}

\item{mean_2}{The mean of group 2}

\item{s_1}{The standard deviation of group 1 (i.e., the square root of the unbiased estimator of the population variance)}

\item{s_2}{The standard deviation of group 2 (i.e., the square root of the unbiased estimator of the population variance)}

\item{s}{The pooled group standard deviation (i.e., the square root of the unbiased estimator of the population variance)}

\item{n_1}{The sample size within group 1}

\item{n_2}{The sample size within group 2}

\item{unbiased}{Returns the unbiased estimate of the standardized mean difference}
}
\value{
A 1-row \code{data.frame} with columns \code{term} (\code{"smd"}) and
\code{value} (the estimated standardized mean difference).
}
\description{
Estimates the standardized mean difference (Cohen's \emph{d}), the
difference between two group means divided by the pooled standard
deviation, from either raw data or summary statistics. Expressing the
difference in standard deviation units frees the comparison from the raw
measurement units, so effects can be compared across measures and studies;
either the ordinary or the unbiased (Hedges, 1981) estimate can be
returned.
}
\details{
When \code{unbiased=TRUE}, the unbiased estimate of the standardized mean difference is returned (Hedges, 1981).
}
\examples{
# Generate sample data.
set.seed(113)
g.1 <- rnorm(n = 25, mean = .5, sd = 1)
g.2 <- rnorm(n = 25, mean = 0, sd = 1)
smd(group_1 = g.1, group_2 = g.2)

M.x <- .66745
M.y <- .24878
sd <- 1.048
smd(mean_1 = M.x, mean_2 = M.y, s = sd)

M.x <- .66745
M.y <- .24878
n1 <- 25
n2 <- 25
sd.1 <- .95817
sd.2 <- 1.1311
smd(mean_1 = M.x, mean_2 = M.y, s_1 = sd.1, s_2 = sd.2, n_1 = n1, n_2 = n2)

smd(mean_1 = M.x, mean_2 = M.y, s_1 = sd.1, s_2 = sd.2, n_1 = n1, n_2 = n2,
    unbiased = TRUE)

}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral
  sciences} (2nd ed.). Hillsdale, NJ: Lawrence Erlbaum.

Cumming, G., & Finch, S. (2001). A primer on the understanding, use, and
  calculation of confidence intervals that are based on central and
  noncentral distributions. \emph{Educational and Psychological
  Measurement, 61}(4), 532--574. \doi{10.1177/0013164401614002}

Hedges, L. V. (1981). Distribution theory for Glass's Estimator of effect size and related estimators.
\emph{Journal of Educational Statistics, 6}(2), 107--128.

Kelley, K. (2005) The effects of nonnormal distributions on confidence intervals around the standardized
mean difference: Bootstrap and parametric confidence intervals, \emph{Educational and Psychological
Measurement, 65}, 51--69. \doi{10.1177/0013164404264850}

Kelley, K. (2007). Confidence intervals for standardized effect sizes:
Theory, application, and implementation. \emph{Journal of Statistical Software, 20}(8), 1--24.
  \doi{10.18637/jss.v020.i08}

Kelley, K., & Rausch, J. R. (2006). Sample size planning for the
  standardized mean difference: Accuracy in parameter estimation via
  narrow confidence intervals. \emph{Psychological Methods, 11}(4),
  363--385. \doi{10.1037/1082-989X.11.4.363}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing
  experiments and analyzing data: A model comparison perspective}
  (4th ed.). Routledge. (See Chapter 4 on individual comparisons and
  Chapter 3 on one-way ANOVA.)

Steiger, J. H., & Fouladi, R. T. (1997). Noncentrality interval
  estimation and the evaluation of statistical methods. In L. L. Harlow,
  S. A. Mulaik, & J. H. Steiger (Eds.), \emph{What if there were no
  significance tests?} (pp. 221--257). Mahwah, NJ: Lawrence Erlbaum.
}
\seealso{
\code{\link{smd_c}}, \code{\link{ci_smd}}, \code{\link{ci_smd_c}},
\code{\link{ss_aipe_smd}}, \code{\link{ss_power_smd}},
\code{\link{plot_smd}}, \code{\link{ci_nc_t}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\concept{Cohen's d}
\concept{Hedges' g}
\keyword{htest}
\keyword{models}
