% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterdata.R
\name{splitData}
\alias{splitData}
\title{split data}
\usage{
splitData(data, p = c(0.8, 0.2))
}
\arguments{
\item{data}{a dataframe of dataset}

\item{p}{split probability}
}
\value{
a list of split dataset
}
\description{
split data
}
\examples{
splitData(iris)
}
