% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_aipe_rmsea.R
\name{ss_aipe_rmsea}
\alias{ss_aipe_rmsea}
\title{Sample Size Planning for RMSEA in SEM}
\usage{
ss_aipe_rmsea(RMSEA, df, width, conf_level = 0.95)
}
\arguments{
\item{RMSEA}{The input RMSEA value}

\item{df}{Degrees of freedom of the model}

\item{width}{Desired confidence interval width}

\item{conf_level}{Desired confidence level (e.g., .90, .95, .99, etc.)}
}
\value{
Returns the necessary total sample size in order to achieve the desired degree of accuracy
(i.e., the sufficiently narrow confidence interval).
}
\description{
Sample size planning for the population root mean square error of approximation (RMSEA) from the accuracy
in parameter estimation (AIPE) perspective. The sample size is planned so that the expected width of a
confidence interval for the population RMSEA is no larger than desired.
}
\examples{
ss_aipe_rmsea(RMSEA = .035, df = 50, width = .05, conf_level = .95)

}
\references{
Kelley, K., & Lai, K. (2011). Accuracy in parameter estimation for the
  root mean square error of approximation: Sample size planning for
  narrow confidence intervals.
  \emph{Multivariate Behavioral Research, 46}, 1--32.
  \doi{10.1080/00273171.2011.543027}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing
  experiments and analyzing data: A model comparison perspective}
  (4th ed.). Routledge.
}
\seealso{
\code{\link{ci_rmsea}}

\code{\link{design_consequences}} for what a chosen design delivers:
  power, the Type S (sign) and Type M (exaggeration) errors of the
  significance filter, and the expected confidence interval width.
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{design}
