% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{SubtitleTheme}
\alias{SubtitleTheme}
\title{Set Subtitle theme}
\usage{
SubtitleTheme(
  ploSubtitleFace = NULL,
  ploSubtitleColor = NULL,
  ploSubtitleSize = NULL,
  ploSubtitleHjust = NULL,
  ploSubtitleVjust = NULL,
  ploSubtitleAngle = 0
)
}
\arguments{
\item{ploSubtitleFace}{subtitle face}

\item{ploSubtitleColor}{subtitle color}

\item{ploSubtitleSize}{subtitle size}

\item{ploSubtitleHjust}{subtitle horizontal position}

\item{ploSubtitleVjust}{subtitle vertical position}

\item{ploSubtitleAngle}{subtitle angle}
}
\value{
A ggolot2 object
}
\description{
Set Subtitle theme
}
