% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_k.R
\name{tidy.dmar_cfa_k}
\alias{tidy.dmar_cfa_k}
\title{Tidy a Multiple-Factor CFA Fit}
\usage{
\method{tidy}{dmar_cfa_k}(x, ...)
}
\arguments{
\item{x}{A \code{dmar_cfa_k} object returned by \code{\link{cfa_k}}.}

\item{\dots}{Unused.}
}
\value{
A \code{data.frame} with columns \code{term},
  \code{estimate}, \code{se}, \code{statistic},
  \code{p_value}, \code{ci_lower}, \code{ci_upper}.
}
\description{
Returns the parameter rows of a \code{\link{cfa_k}} table (loadings,
error variances, intercepts, latent correlations, and the defined
measurement properties) in the column convention used by the
\pkg{broom} ecosystem. Fit-information rows belong in
\code{\link{glance.dmar_cfa_k}}.
}
\examples{
data(holzinger_swineford)
res <- cfa_k(holzinger_swineford,
             list(verbal = c("t6_paragraph_comprehension",
                             "t7_sentence", "t9_word_meaning"),
                  deduction = c("t20_deduction",
                                "t22_problem_reasoning",
                                "t23_series_completion")))
generics::tidy(res)
generics::glance(res)

}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
