% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{tidy.dmar_reliability}
\alias{tidy.dmar_reliability}
\title{A Reliability Coefficient Estimate}
\usage{
\method{tidy}{dmar_reliability}(x, ...)
}
\arguments{
\item{x}{A \code{dmar_reliability} object returned by any of
\code{\link{reliability_alpha}}, \code{\link{reliability_kr20}},
\code{\link{reliability_omega}},
\code{\link{reliability_omega_categorical}}, or \code{\link{reliability}}.}

\item{\dots}{Unused.}
}
\value{
A one-row \code{data.frame}.
}
\description{
Returns a one-row \code{data.frame} in the column convention
used by the \pkg{broom} ecosystem (\code{term},
\code{estimate}, \code{se}, \code{ci_lower},
\code{ci_upper}). The \code{term} is the coefficient name
(\code{"alpha"}, \code{"omega"}, etc.).
}
\examples{
# Coefficient alpha for the three verbal tests of the Holzinger and
# Swineford battery, from their covariance matrix.
S <- cov(holzinger_swineford[, c("t6_paragraph_comprehension",
                                 "t7_sentence", "t9_word_meaning")])
res <- reliability_alpha(S = S, N = 301, ci_method = "feldt")
generics::tidy(res)
generics::glance(res)

}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
