% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{TiitleTheme}
\alias{TiitleTheme}
\title{Set Title theme}
\usage{
TiitleTheme(
  ploTitleFace = NULL,
  ploTitleColor = NULL,
  ploTitleSize = NULL,
  ploTitleHjust = NULL,
  ploTitleVjust = NULL,
  ploTitleAngle = NULL
)
}
\arguments{
\item{ploTitleFace}{title face}

\item{ploTitleColor}{title color}

\item{ploTitleSize}{tilte size}

\item{ploTitleHjust}{title horizontal position}

\item{ploTitleVjust}{title vertical position}

\item{ploTitleAngle}{title angle}
}
\value{
A ggolot2 object
}
\description{
Set Title theme
}
