% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{TitleSetting}
\alias{TitleSetting}
\title{Title Setting}
\usage{
TitleSetting(
  titleText = "标题",
  subtitleText = "副标题",
  captionText = "说明文字",
  xlabText = "x轴线标签",
  ylabText = "y轴线标签"
)
}
\arguments{
\item{titleText}{title text}

\item{subtitleText}{subtitle text}

\item{captionText}{caption text}

\item{xlabText}{xlab text}

\item{ylabText}{ylab text}
}
\value{
A ggolot2 object
}
\description{
Title Setting
}
