% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_R2.R
\name{var_R2}
\alias{var_R2}
\title{Variance of the Squared Multiple Correlation Coefficient}
\usage{
var_R2(population_R2, N, p)
}
\arguments{
\item{population_R2}{Population squared multiple correlation coefficient}

\item{N}{Sample size}

\item{p}{The number of predictor variables}
}
\value{
A 1-row \code{data.frame} with columns \code{term} and \code{value}.
The \code{term} value is \code{"var_R2"} and \code{value} is the
asymptotic variance of \eqn{R^2}.
}
\description{
Computes the sampling variance of the squared multiple correlation
coefficient from the population value, the sample size, and the
number of predictors, the quantity that governs how precisely
\eqn{R^2} is estimated at a given design size.
}
\details{
Uses the hypergeometric function as discussed in and section 28 of Stuart, Ord, and Arnold (1999) in
order to obtain the \emph{correct} value for the variance of the squared multiple correlation coefficient.
}
\note{
The Gauss hypergeometric function \eqn{{}_2F_1} is computed in base R (see
the internal \code{.hyperg_2F1}); no GSL system library is required.
}
\examples{
var_R2(.5, 10, 5)
var_R2(.5, 25, 5)
var_R2(.5, 50, 5)
var_R2(.5, 100, 5)

}
\references{
Kelley, K. (2008). Sample size planning for the squared multiple
  correlation coefficient: Accuracy in parameter estimation via narrow
  confidence intervals. \emph{Multivariate Behavioral Research, 43},
  524--555. \doi{10.1080/00273170802490632}

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027). \emph{Designing
  experiments and analyzing data: A model comparison perspective}
  (4th ed.). Routledge. (See Chapter 3 on \eqn{R^2} as a model comparison
  effect size.)

Stuart, A., Ord, J. K., & Arnold, S. (1999). \emph{Kendall's advanced
  theory of statistics, volume 2A: Classical inference and the linear
  model} (6th ed.). Arnold.
}
\seealso{
\code{\link{expected_R2}}, \code{\link{ci_R2}}, \code{\link{ss_aipe_R2}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\keyword{design}
