% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{XlabTheme}
\alias{XlabTheme}
\title{Axis x}
\usage{
XlabTheme(
  xlabAxisTitleTextFace = NULL,
  xlabAxisTitleTextColor = NULL,
  xlabAxisTitleTextSize = NULL,
  xlabAxisTitleTextHjust = NULL,
  xlabAxisTitleTextVjust = NULL,
  xlabAxisTitleTextAngle = NULL,
  xlabAxisTextFace = NULL,
  xlabAxisTextColor = NULL,
  xlabAxisTextSize = NULL,
  xlabAxisTextHjust = NULL,
  xlabAxisTextVjust = NULL,
  xlabAxisTextAngle = NULL,
  xlabAxisTicksSize = NULL,
  xlabAxisTicksColor = NULL,
  xlabAxisTicksLength = 0.1,
  xlabAxisLineSize = NULL,
  xlabAxisLineColor = NULL,
  xlabAxisLineType = NULL
)
}
\arguments{
\item{xlabAxisTitleTextFace}{x axis title text face}

\item{xlabAxisTitleTextColor}{color}

\item{xlabAxisTitleTextSize}{size}

\item{xlabAxisTitleTextHjust}{x postion}

\item{xlabAxisTitleTextVjust}{y position}

\item{xlabAxisTitleTextAngle}{angle}

\item{xlabAxisTextFace}{x axis text face}

\item{xlabAxisTextColor}{color}

\item{xlabAxisTextSize}{size}

\item{xlabAxisTextHjust}{x postion}

\item{xlabAxisTextVjust}{y postion}

\item{xlabAxisTextAngle}{angle}

\item{xlabAxisTicksSize}{ticks size}

\item{xlabAxisTicksColor}{color}

\item{xlabAxisTicksLength}{length}

\item{xlabAxisLineSize}{line size}

\item{xlabAxisLineColor}{color}

\item{xlabAxisLineType}{type}
}
\value{
A ggolot2 object
}
\description{
Axis x
}
