% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{YlabTheme}
\alias{YlabTheme}
\title{Axis y}
\usage{
YlabTheme(
  ylabAxisTitleTextFace = NULL,
  ylabAxisTitleTextColor = NULL,
  ylabAxisTitleTextSize = NULL,
  ylabAxisTitleTextHjust = NULL,
  ylabAxisTitleTextVjust = NULL,
  ylabAxisTitleTextAngle = NULL,
  ylabAxisTextFace = NULL,
  ylabAxisTextColor = NULL,
  ylabAxisTextSize = NULL,
  ylabAxisTextHjust = NULL,
  ylabAxisTextVjust = NULL,
  ylabAxisTextAngle = NULL,
  ylabAxisTicksSize = NULL,
  ylabAxisTicksColor = NULL,
  ylabAxisTicksLength = 0.1,
  ylabAxisLineSize = NULL,
  ylabAxisLineColor = NULL,
  ylabAxisLineType = NULL
)
}
\arguments{
\item{ylabAxisTitleTextFace}{axis title text face}

\item{ylabAxisTitleTextColor}{color}

\item{ylabAxisTitleTextSize}{size}

\item{ylabAxisTitleTextHjust}{x postion}

\item{ylabAxisTitleTextVjust}{y position}

\item{ylabAxisTitleTextAngle}{angle}

\item{ylabAxisTextFace}{axis text face}

\item{ylabAxisTextColor}{color}

\item{ylabAxisTextSize}{size}

\item{ylabAxisTextHjust}{x postion}

\item{ylabAxisTextVjust}{y position}

\item{ylabAxisTextAngle}{angle}

\item{ylabAxisTicksSize}{ticks size}

\item{ylabAxisTicksColor}{color}

\item{ylabAxisTicksLength}{length}

\item{ylabAxisLineSize}{line size}

\item{ylabAxisLineColor}{color}

\item{ylabAxisLineType}{type}
}
\value{
A ggolot2 object
}
\description{
Axis y
}
