% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{Ylim}
\alias{Ylim}
\title{Axis y range}
\usage{
Ylim(start = NULL, end = NULL)
}
\arguments{
\item{start}{start point of y axis}

\item{end}{end point of y axis}
}
\value{
A ggplot2 object <-
}
\description{
Axis y range
}
