% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_sm.R
\name{ci_sm}
\alias{ci_sm}
\title{Confidence Interval for the Standardized Mean}
\usage{
ci_sm(
  sm = NULL,
  mean = NULL,
  sd = NULL,
  ncp = NULL,
  N = NULL,
  conf_level = 0.95,
  alpha_lower = NULL,
  alpha_upper = NULL,
  ...
)
}
\arguments{
\item{sm}{Standardized mean}

\item{mean}{Mean}

\item{sd}{Standard deviation}

\item{ncp}{Noncentral parameter}

\item{N}{Sample size}

\item{conf_level}{Confidence interval coverage (i.e., 1 - Type I error rate); default is .95}

\item{alpha_lower}{Type I error for the lower confidence limit}

\item{alpha_upper}{Type I error for the upper confidence limit}

\item{\dots}{Allows one to potentially include parameter values for inner functions}
}
\value{
A 3-row \code{data.frame} with columns \code{term} and \code{value}. The
\code{term} values are \code{"lower_limit"} (the lower confidence limit
on the standardized mean), \code{"std_mean"} (the standardized mean),
and \code{"upper_limit"} (the upper confidence limit).
}
\description{
Computes the exact confidence interval for the standardized mean,
the mean divided by the standard deviation, by inverting the
noncentral \emph{t} distribution. The standardized mean is the
one-sample analog of the standardized mean difference and shares
its noncentral interval theory.
}
\details{
The user must specify the standardized mean in one and only one of the three ways:
a) mean and standard deviation (\code{mean} and \code{sd}),
b) standardized mean (\code{sm}), and
c) noncentral parameter (\code{ncp}).
The confidence level must be specified in one of following two ways: using confidence interval coverage (\code{conf_level}),
or lower and upper confidence limits (\code{alpha_lower} and \code{alpha_upper}). This function uses the exact confidence
interval method based on noncentral \emph{t}-distributions. The confidence interval for noncentral \emph{t}-parameter can
be obtained from the \code{ci_nc_t} function in DMAR.
}
\note{
The standardized mean is the mean divided by the standard deviation.
}
\examples{
ci_sm(sm = 2.037905, N = 13, conf_level = .95)
ci_sm(mean = 30, sd = 14.721, N = 13, conf_level = .95)
ci_sm(ncp = 7.347771, N = 13, conf_level = .95)
ci_sm(sm = 2.037905, N = 13, alpha_lower = .05, alpha_upper = 0)
ci_sm(mean = 50, sd = 10, N = 25, conf_level = .95)

}
\references{
Kelley, K. (2007). Confidence intervals for standardized effect sizes:
Theory, application, and implementation. \emph{Journal of Statistical Software, 20}(8), 1--24.
  \doi{10.18637/jss.v020.i08}

Steiger, J. H., & Fouladi, R. T. (1997). Noncentrality interval
  estimation and the evaluation of statistical methods. In L. L. Harlow,
  S. A. Mulaik, & J. H. Steiger (Eds.), \emph{What if there were no
  significance tests?} (pp. 221--257). Mahwah, NJ: Lawrence Erlbaum.
}
\seealso{
\code{\link{ci_nc_t}}

Other confidence intervals for effect sizes:
\code{\link[=ci_R2]{ci_R2()}},
\code{\link[=ci_c]{ci_c()}},
\code{\link[=ci_c_ancova]{ci_c_ancova()}},
\code{\link[=ci_c_ancova_bp]{ci_c_ancova_bp()}},
\code{\link{ci_correlation}},
\code{\link[=ci_cv]{ci_cv()}},
\code{\link[=ci_eta_squared]{ci_eta_squared()}},
\code{\link[=ci_eta_squared_generalized]{ci_eta_squared_generalized()}},
\code{\link[=ci_eta_squared_partial]{ci_eta_squared_partial()}},
\code{\link[=ci_mahalanobis]{ci_mahalanobis()}},
\code{\link[=ci_omega_squared]{ci_omega_squared()}},
\code{\link[=ci_pvaf]{ci_pvaf()}},
\code{\link[=ci_rc]{ci_rc()}},
\code{\link[=ci_reg_coef]{ci_reg_coef()}},
\code{\link[=ci_rmsea]{ci_rmsea()}},
\code{\link[=ci_sc]{ci_sc()}},
\code{\link[=ci_sc_ancova]{ci_sc_ancova()}},
\code{\link[=ci_smd]{ci_smd()}},
\code{\link[=ci_smd_c]{ci_smd_c()}},
\code{\link[=ci_snr]{ci_snr()}},
\code{\link[=ci_src]{ci_src()}},
\code{\link[=ci_srsnr]{ci_srsnr()}},
\code{\link[=contrast_adjusted]{contrast_adjusted()}},
\code{\link[=plot_smd]{plot_smd()}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\concept{confidence intervals for effect sizes}
\keyword{design}
