% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_src.R
\name{ci_src}
\alias{ci_src}
\title{Confidence Interval for a Standardized Regression Coefficient}
\usage{
ci_src(
  beta_j = NULL,
  SE_beta_j = NULL,
  N = NULL,
  p = NULL,
  R2_Y_X = NULL,
  R2_j_X_without_j = NULL,
  conf_level = 0.95,
  R2_Y_X_without_j = NULL,
  t_value = NULL,
  b_j = NULL,
  SE_b_j = NULL,
  s_Y = NULL,
  s_X = NULL,
  alpha_lower = NULL,
  alpha_upper = NULL,
  ...
)
}
\arguments{
\item{beta_j}{The standardized regression coefficient}

\item{SE_beta_j}{The standard error of the standardized regression coefficient}

\item{N}{Sample size}

\item{p}{The number of predictors}

\item{R2_Y_X}{The squared multiple correlation coefficient predicting \emph{Y} from the \emph{p} predictor variables}

\item{R2_j_X_without_j}{The squared multiple correlation coefficient predicting the \emph{j}th predictor variable (i.e., the predictor of interest) from the remaining \emph{p}-1 predictor variables}

\item{conf_level}{Desired level of confidence for the computed interval (i.e., 1 - the Type I error rate)}

\item{R2_Y_X_without_j}{The squared multiple correlation coefficient predicting \emph{Y} from the \emph{p}-1 predictor variable with the \emph{j}th predictor of interest excluded}

\item{t_value}{The \emph{t}-value evaluating the null hypothesis that the population regression coefficient for the \emph{j}th predictor equals zero}

\item{b_j}{The unstandardized regression coefficient}

\item{SE_b_j}{The standard error of the unstandardized regression coefficient}

\item{s_Y}{Standard deviation of \emph{Y}, the dependent variable}

\item{s_X}{Standard deviation of \emph{X}, the predictor variable of interest}

\item{alpha_lower}{The Type I error rate for the lower confidence interval limit}

\item{alpha_upper}{The Type I error rate for the upper confidence interval limit}

\item{\dots}{Optional additional specifications for nested functions}
}
\value{
A 3-row \code{data.frame} with columns \code{term}, \code{value},
\code{prob_less}, and \code{prob_greater}. The \code{term} values are
\code{"lower_limit"}, \code{"src"} (the standardized regression
coefficient point estimate), and \code{"upper_limit"}, so the estimate
sits between its confidence limits. The \code{prob_less} and
\code{prob_greater} columns report the achieved tail probabilities at
each limit when the noncentral t method is used (\code{NA} for the
estimate row).
}
\description{
Computes a confidence interval for a population standardized
regression coefficient, by the standard \emph{t}-based approach or
the noncentral \emph{t} approach. A thin convenience wrapper around
\code{\link{ci_reg_coef}}, which is the general engine; for the
coefficient in its raw metric use \code{\link{ci_rc}}.
}
\details{
For standardized variables, do not specify the standard deviation of the variables and input the
standardized regression coefficient for \code{b_j}.
}
\note{
This function calls upon \code{ci_reg_coef} in DMAR, but has a different naming scheme.
See  \code{\link{ci_reg_coef}} for more details.

To form a confidence interval for the unstandardized regression coefficient, use \code{ci_rc}.
This function is used to form a confidence interval for the standardized regression coefficient.

Not all of the values need to be specified, only those that contain all of the necessary information
in order to compute the confidence interval (options are thus given for the values that need to be specified).
}
\examples{
ci_src(beta_j = .6707, .1761, N = 30, p = 6, conf_level = .95)

}
\references{
Kelley, K. (2007). Confidence intervals for standardized effect sizes:
Theory, application, and implementation. \emph{Journal of Statistical Software, 20}(8), 1--24.
  \doi{10.18637/jss.v020.i08}

Kelley, K., & Maxwell, S. E. (2003). Sample size for multiple
  regression: Obtaining regression coefficients that are accurate,
  not simply significant. \emph{Psychological Methods, 8}(3),
  305--321. \doi{10.1037/1082-989X.8.3.305}

Kelley, K., & Maxwell, S. E. (2008). Sample size planning with
  applications to multiple regression: Power and accuracy for omnibus
  and targeted effects. In P. Alasuutari, L. Bickman, & J. Brannen
  (Eds.), \emph{The Sage handbook of social research methods}
  (pp. 166--192). Sage.

Maxwell, S. E., Delaney, H. D., & Kelley, K. (2027).
  \emph{Designing experiments and analyzing data: A model comparison
  perspective} (4th ed.). Routledge. (See Chapter 4 on individual
  comparisons of means and Chapter 6 on trend analysis.)

Smithson, M. (2003). \emph{Confidence intervals}. Thousand Oaks, CA: Sage Publications.

Steiger, J. H. (2004). Beyond the \emph{F} Test: Effect size confidence intervals and tests of close
fit in the Analysis of Variance and Contrast Analysis. \emph{Psychological Methods, 9}(2), 164--182.
  \doi{10.1037/1082-989X.9.2.164}
}
\seealso{
\code{\link{ss_aipe_reg_coef}}, \code{\link{conf_limits_nct}}, \code{\link{ci_reg_coef}}, \code{\link{ci_rc}}

Other confidence intervals for effect sizes:
\code{\link[=ci_R2]{ci_R2()}},
\code{\link[=ci_c]{ci_c()}},
\code{\link[=ci_c_ancova]{ci_c_ancova()}},
\code{\link[=ci_c_ancova_bp]{ci_c_ancova_bp()}},
\code{\link{ci_correlation}},
\code{\link[=ci_cv]{ci_cv()}},
\code{\link[=ci_eta_squared]{ci_eta_squared()}},
\code{\link[=ci_eta_squared_generalized]{ci_eta_squared_generalized()}},
\code{\link[=ci_eta_squared_partial]{ci_eta_squared_partial()}},
\code{\link[=ci_mahalanobis]{ci_mahalanobis()}},
\code{\link[=ci_omega_squared]{ci_omega_squared()}},
\code{\link[=ci_pvaf]{ci_pvaf()}},
\code{\link[=ci_rc]{ci_rc()}},
\code{\link[=ci_reg_coef]{ci_reg_coef()}},
\code{\link[=ci_rmsea]{ci_rmsea()}},
\code{\link[=ci_sc]{ci_sc()}},
\code{\link[=ci_sc_ancova]{ci_sc_ancova()}},
\code{\link[=ci_sm]{ci_sm()}},
\code{\link[=ci_smd]{ci_smd()}},
\code{\link[=ci_smd_c]{ci_smd_c()}},
\code{\link[=ci_snr]{ci_snr()}},
\code{\link[=ci_srsnr]{ci_srsnr()}},
\code{\link[=contrast_adjusted]{contrast_adjusted()}},
\code{\link[=plot_smd]{plot_smd()}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\concept{confidence intervals for effect sizes}
\keyword{design}
