% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_R2.R
\name{convert_R2}
\alias{convert_R2}
\alias{convert_R2_f}
\alias{convert_f_R2}
\alias{convert_lambda_R2}
\alias{convert_R2_lambda}
\title{Convert Between \emph{F}, \eqn{R^2}, and Their Noncentral
  Parameters}
\usage{
convert_R2_f(R2 = NULL, df_1 = NULL, df_2 = NULL, p = NULL, N = NULL)

convert_f_R2(F_value = NULL, df_1 = NULL, df_2 = NULL)

convert_lambda_R2(lambda = NULL, N = NULL)

convert_R2_lambda(R2 = NULL, N = NULL)
}
\arguments{
\item{R2}{Squared multiple correlation coefficient (population or observed)}

\item{df_1}{Degrees of freedom for the numerator of the \emph{F}-distribution}

\item{df_2}{Degrees of freedom for the denominator of the \emph{F}-distribution}

\item{p}{Number of predictor variables for \code{R2}}

\item{N}{Sample size}

\item{F_value}{The obtained \emph{F} value from a test of significance for the squared multiple correlation coefficient}

\item{lambda}{The noncentral parameter from an \emph{F}-distribution}
}
\value{
Each of the four functions returns a 1-row \code{data.frame}
  with columns \code{term} and \code{value}.  The \code{term} entry
  identifies the conversion performed
  (\code{"r2_f"}, \code{"f_r2"}, \code{"lambda_r2"}, or
  \code{"r2_lambda"}) and \code{value} is the converted scalar. The
  conversions are exact inverses of one another (with the appropriate
  degrees-of-freedom / sample size inputs supplied), which is what
  makes them useful inside the noncentrality-parameter confidence
  interval machinery of \code{\link{ci_R2}} and
  \code{\link{ci_nc_F}}.
}
\description{
Given values of test statistics (and the appropriate additional information) the value of the noncentral
values can be obtained. Likewise, given noncentral values (and the appropriate additional information)
the value of the test statistic can be obtained.
}
\details{
These functions are especially helpful in the search for confidence intervals for noncentral parameters, as they convert to and from related quantities.
}
\examples{
convert_R2_lambda(R2 = .5, N = 100)

}
\references{
Kelley, K. (2007). Confidence intervals for standardized effect sizes:
Theory, application, and implementation. \emph{Journal of Statistical
Software, 20}(8), 1--24. \doi{10.18637/jss.v020.i08}
}
\seealso{
\code{\link{ss_aipe_R2}}, \code{\link{ci_R2}}, \code{\link{ci_nc_t}}, \code{\link{ci_nc_F}}

Other parameterization conversions:
\code{\link[=convert_F_chisq]{convert_F_chisq()}},
\code{\link[=convert_Z_r]{convert_Z_r()}},
\code{\link[=convert_cor_cov]{convert_cor_cov()}},
\code{\link[=convert_d_or]{convert_d_or()}},
\code{\link[=convert_d_r]{convert_d_r()}},
\code{\link[=convert_r_Z]{convert_r_Z()}},
\code{\link{convert_t_smd}},
\code{\link[=convert_z_normal]{convert_z_normal()}}
}
\author{
Ken Kelley \email{kkelley@nd.edu}
}
\concept{parameterization conversions}
\keyword{design}
\keyword{multivariate}
