% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterdata.R
\name{filterFun}
\alias{filterFun}
\title{filter data with cor,miss, nzv}
\usage{
filterFun(
  formula = formula(NULL),
  data,
  corthreshold = 0.9,
  method = "pearson",
  misthreshold = 0.1
)
}
\arguments{
\item{formula}{formula of data}

\item{data}{data set, a dataframe}

\item{corthreshold}{threshold of corr}

\item{method}{method of corr}

\item{misthreshold}{threshold of miss value proportion}
}
\value{
A recipe object
}
\description{
filter data with cor,miss, nzv
}
\examples{
data(credit_data, package = "modeldata")
result <- filterFun(Status ~ ., data = credit_data,data = credit_data)
bake(result,credit_data)
}
