% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{TitleSetting}
\alias{TitleSetting}
\title{Title Setting}
\usage{
TitleSetting(
  titleText = NULL,
  subtitleText = NULL,
  captionText = NULL,
  xlabText = NULL,
  ylabText = NULL
)
}
\arguments{
\item{titleText}{title text}

\item{subtitleText}{subtitle text}

\item{captionText}{caption text}

\item{xlabText}{xlab text}

\item{ylabText}{ylab text}
}
\value{
A ggolot2 object
}
\description{
Title Setting
}
