% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV.R
\name{Xlim}
\alias{Xlim}
\title{Axis x range}
\usage{
Xlim(start = NULL, end = NULL)
}
\arguments{
\item{start}{start point of x axis}

\item{end}{point of x axis}
}
\value{
A ggplot2 object
}
\description{
Axis x range
}
